src/test/java/com/leetcode/algorithm/easy/palindromelinkedlist/SolutionTest.java
package com.leetcode.algorithm.easy.palindromelinkedlist;

import static org.junit.jupiter.api.Assertions.assertEquals;

import com.leetcode.algorithm.common.ListNode;
import org.junit.jupiter.api.Test;

public class SolutionTest {
  @Test
  public void testCase1() {
    Solution solution = new Solution();
    ListNode head = new ListNode(1);
    head.next = new ListNode(2);
    assertEquals(false, solution.isPalindrome(head));
  }

  @Test
  public void testCase2() {
    Solution solution = new Solution();
    ListNode head = new ListNode(1);
    head.next = new ListNode(2);
    head.next.next = new ListNode(2);
    head.next.next.next = new ListNode(1);
    assertEquals(true, solution.isPalindrome(head));
  }
}
